ti_config_fragments: rpmsg/v8_rpmsg: Enable PRUSS support
Add support to build all the PRUSS platform drivers - PRUSS driver, the
PRUSS INTC IrqChip driver and the PRU remoteproc driver as modules by
default. This will enable the PRU-ICSS instance(s) present on various
AM33xx, AM437x, AM57xx, 66AK2G, and the K3 AM65x and J721E family of
SoCs.

Note that the PRUSS INTC module is a silent Kconfig option, and so
need not be enabled explicitly, but is being added for conformity
with older kernels.

Signed-off-by: Suman Anna <s-anna@ti.com>