arm64: dts: ti: k3-am62-main: Add node for Display SubSystem
Add DT node for the Display SubSystem on the am62x soc in cbass_main.
The DSS IP on this soc is compatible with the one on the am65x soc.

The DSS supports one each of video pipeline (vid) and video-lite
pipeline (vidl1). It outputs OLDI signals on one video port (vp1) and
DPI signals on another (vp2). The video ports are connected to the
pipelines via 2 identical overlay managers (ovr1 and ovr2).

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Tested-by: Darren Etheridge <detheridge@ti.com>