net: ti: icssg-prueth: move phy_connect()/phy_disconnect()
The phy_connect() and phy_disconnect() functions are currently called
when the interface is brought up and down using ifconfig respectively.
This causes an issue when ethtool is used after bringing down the
interface. This issue was reported by Siemens.

Move the phy_connect()/phy_disconnect() logic to probe and remove
respectively to solve this bug.

Signed-off-by: Puranjay Mohan <p-mohan@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Link: https://serenity.dal.design.ti.com/lore/linux-patch-review/20220509053931.4552-1-p-mohan@ti.com