phy: ti: gmii-sel: Add support for CPSW9G GMII SEL in J721E
Each of the CPSW9G ports in J721E support additional modes like QSGMII.
Add a new compatible for J721E to support the additional modes.

In TI's J721E, each of the CPSW9G ethernet interfaces can act as a
QSGMII or QSGMII-SUB port. The QSGMII interface is responsible for
performing auto-negotiation between the MAC and the PHY while the rest of
the interfaces are designated as QSGMII-SUB interfaces, indicating that
they will not be taking part in the auto-negotiation process.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>