net: ethernet: ti: am65-cpsw: Add RMII mode to supported interfaces
Support for RMII mode is already present in the am65-cpsw driver.
Currently, RMII mode is not set as a supported phy_mode in struct
"phylink_config"'s supported_interfaces member. This prevents an interface
from being used in RMII mode.

Fix it by adding RMII mode as a supported phy-mode if requested by the
interface.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>